Sign-in working properly but pages slightly buggy
Sign-in is working properly. You can sign in as a customer or as a
restaurant, with customers being able to make reservations based on
their cID and the restaurant's rID. Restaurants and people not logged in
are unable to make a reservation and will instead be sent to the sign-in
page.

There is a bug with displaying the "log out" button that i've put on the
nav bar. I've added it to all pages, but it does not show up on the
sign-in or home pages unless you are sent there by a method other than
using the nav bar (ex logging out, logging in, making a reservation
etc.). if you navigate to one of these pages via the nav bar, this
button does not appear. This will be fixed in a future commit.